refac: Moved data classes to a data file
